+++ src/models/uniqueTable.hpp
@@ -9,18 +9,18 @@
 using namespace std;
 
 struct Entry {
-    size_t key;
+    long long key;
     shared_ptr<QMDDNode> value;
-    Entry(size_t k, shared_ptr<QMDDNode> v) : key(k), value(v) {}
+    Entry(long long k, shared_ptr<QMDDNode> v) : key(k), value(v) {}
 };
 
 class UniqueTable {
 private:
-    unordered_map<size_t, vector<Entry>> table;
+    unordered_map<long long, vector<Entry>> table;
     mutable shared_mutex tableMutex;
     const size_t tableSize ;
     UniqueTable();
-    size_t hash(size_t hashKey) const;
+    long long hash(long long hashKey) const;
 
 
 
@@ -28,8 +28,8 @@ public:
     UniqueTable(const UniqueTable&) = delete;
     UniqueTable& operator=(const UniqueTable&) = delete;
     static UniqueTable& getInstance();
-    void insert(size_t hashKey, shared_ptr<QMDDNode> node);
-    shared_ptr<QMDDNode> find(size_t hashKey) const;
+    void insert(long long hashKey, shared_ptr<QMDDNode> node);
+    shared_ptr<QMDDNode> find(long long hashKey) const;
     void printAllEntries() const;
 };
